Added compass heading
At the moment, speed will not be calculated if we always have a compass reading available, this could mean we don't get anything telemetry wise at low speed. I am also a little wary that if the speed loop is too slow we could have error associated with taking an out of date compass reading.